Move the full workspace to ES Module
