+++ .changeset/tame-news-live.md
@@ -1,0 +1,7 @@
+---
+"eslint-config-custom": major
+"usehooks-ts": major
+"www": major
+---
+
+Move the full workspace into ES Module
