+++ apps/www/next.config.js
@@ -1,4 +1,4 @@
-import './env.mjs'
+import './env.js'
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {}
