+++ package.json
@@ -5,6 +5,7 @@
   "author": "Julien CARON <juliencaron@protonmail.com>",
   "homepage": "https://usehooks-ts.com",
   "module": "true",
+  "type": "module",
   "keywords": [
     "typescript",
     "react",
@@ -19,10 +20,10 @@
     "clean": "rimraf .turbo && turbo run clean",
     "format": "prettier --write \"**/*.{json,md,mdx,css,scss,yaml,yml}\"",
     "lint": "turbo run lint",
-    "update-testing-issue": "zx ./scripts/updateTestingIssue.mjs",
-    "update-readme": "zx ./scripts/updateReadme.mjs",
-    "copy-hooks": "rimraf -rf ./apps/www/generated && zx ./scripts/copyHooks.mjs",
-    "gen-hook": "turbo gen hook && pnpm format",
+    "update-testing-issue": "zx ./scripts/updateTestingIssue.js",
+    "update-readme": "zx ./scripts/updateReadme.js",
+    "copy-hooks": "rimraf ./apps/www/generated && zx ./scripts/copyHooks.js",
+    "gen-hook": "turbo gen hook --config \"turbo/generators/config.cts\" && pnpm format",
     "changeset": "npx changeset",
     "changeset-version": "npx changeset version",
     "changeset-publish": "npx changeset publish"
@@ -32,7 +33,9 @@
   },
   "devDependencies": {
     "@changesets/cli": "^2.27.1",
+    "@turbo/gen": "^1.12.4",
     "all-contributors-cli": "^6.26.1",
+    "date-fns": "^3.3.1",
     "eslint": "^8.56.0",
     "prettier": "^3.2.5",
     "rimraf": "^5.0.5",
