+++ packages/eslint-config-custom/.eslintrc.cjs
