+++ packages/eslint-config-custom/index.cjs
@@ -1,0 +1,3 @@
+const eslintrc = require('./.eslintrc.cjs')
+
+module.exports = eslintrc
