--- packages/eslint-config-custom/index.js
@@ -1,3 +1,0 @@
-const eslintrc = require('./.eslintrc.js')
-
-module.exports = eslintrc
