+++ packages/eslint-config-custom/package.json
@@ -3,9 +3,10 @@
   "private": true,
   "version": "1.2.0",
   "description": "Base configuration for Eslint",
-  "main": "index.js",
+  "main": "index.cjs",
   "author": "Julien CARON <juliencaron@protonmail.com>",
   "license": "MIT",
+  "type": "module",
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^7.0.2",
     "@typescript-eslint/parser": "^7.0.2",
