+++ turbo/generators/templates/hook/hook.mdx.hbs
