+++ turbo/generators/templates/hook/hook.ts.hbs
