+++ turbo/generators/templates/hook/index.ts.hbs
